refactor: PostService 책임 분산
