+++ src/main/java/com/gdsc/hack/controller/PostController.java
@@ -5,9 +5,12 @@ import com.gdsc.hack.domain.Post;
 import com.gdsc.hack.dto.request.MapNodeRequestDto;
 import com.gdsc.hack.dto.request.PostRequestDto;
 import com.gdsc.hack.global.dto.ResponseDto;
+import com.gdsc.hack.service.FoodMapService;
+import com.gdsc.hack.service.MapNodeService;
 import com.gdsc.hack.service.PostService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -17,16 +20,18 @@ import java.util.List;
 @RequiredArgsConstructor
 public class PostController {
     private final PostService postService;
+    private final FoodMapService foodMapService;
+    private final MapNodeService mapNodeService;
 
     @PostMapping("/post")
     public ResponseDto<Void> post(
         @RequestBody PostRequestDto requestDto
     ) {
         Post post = postService.writePost(requestDto);
-        FoodMap foodMap = postService.writeFoodMap(post, requestDto);
+        FoodMap foodMap = foodMapService.writeFoodMap(post, requestDto);
         List<MapNodeRequestDto> mapNodeDtoList = requestDto.getFoodMap().getMapNodeList();
 
-        postService.writeMapNode(foodMap, mapNodeDtoList);
+        mapNodeService.writeMapNode(foodMap, mapNodeDtoList);
 
         return ResponseDto.success("저장 완료: 게시글이 성공적으로 작성되었습니다.");
     }
@@ -34,5 +39,7 @@ public class PostController {
 //    @PutMapping("/post")
 //    public ResponseDto<Void> fixPost(
 //        @RequestBody PostRequestDto requestDto
-//    )
+//    ) {
+//        postService
+//    }
 }
