+++ src/main/java/com/gdsc/hack/service/PostService.java
@@ -22,8 +22,6 @@ import java.util.List;
 public class PostService {
     private final PostRepository postRepository;
     private final UserRepository userRepository;
-    private final MapNodeRepository mapNodeRepository;
-    private final FoodMapRepository foodMapRepository;
 
     @Transactional
     public Post writePost(PostRequestDto dto) {
@@ -37,39 +35,6 @@ public class PostService {
                 .build();
 
         return postRepository.save(post);
-    }
-
-    @Transactional
-    public FoodMap writeFoodMap(Post post, PostRequestDto dto) {
-        FoodMap foodMap = FoodMap
-                .builder()
-                .user(post.getUser())
-                .post(post)
-                .build();
-
-        List<MapNodeRequestDto> mapNodeList = dto
-                .getFoodMap()
-                .getMapNodeList();
-        return foodMapRepository.save(foodMap);
-    }
-
-    @Transactional
-    public void writeMapNode(FoodMap foodMap, List<MapNodeRequestDto> dtoList) {
-        dtoList.forEach(nodeDto -> {
-            Double latitude = nodeDto.getLatitude();
-            Double longitude = nodeDto.getLongitude();
-            String restaurantName = nodeDto.getRestaurantName();
-
-            MapNode mapNode = MapNode
-                    .builder()
-                    .name(restaurantName)
-                    .latitude(latitude)
-                    .longitude(longitude)
-                    .foodMap(foodMap)
-                    .build();
-
-            mapNodeRepository.save(mapNode);
-        });
     }
 
 }
